+++ src/__tests__/pages/pupils/index.test.tsx
@@ -3,12 +3,8 @@ import assert from "assert";
 import {
   getStaticPaths,
   getStaticProps,
-  pickAvailableSectionsForLesson,
 } from "@/pages/pupils/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]";
 import * as curriculumApi2023 from "@/node-lib/curriculum-api-2023/__mocks__/index";
-import { allLessonReviewSections } from "@/components/PupilComponents/LessonEngineProvider";
-import pupilLessonOverviewFixture from "@/node-lib/curriculum-api/fixtures/pupilLessonOverview.fixture";
-import { quizQuestions } from "@/node-lib/curriculum-api-2023/fixtures/quizElements.fixture";
 import * as getDownloadResourcesExistence from "@/components/SharedComponents/helpers/downloadAndShareHelpers/getDownloadResourcesExistence";
 
 jest.mock("@/utils/handleTranscript");
@@ -17,42 +13,6 @@ jest.mock(
 );
 
 describe("pages/pupils/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug]/index", () => {
-  describe("pickAvailableSectionsForLesson", () => {
-    it("returns all sections if all are available", () => {
-      const sections = pickAvailableSectionsForLesson(
-        pupilLessonOverviewFixture({
-          starterQuiz: quizQuestions,
-          exitQuiz: quizQuestions,
-          videoMuxPlaybackId: "123",
-        }),
-      );
-
-      expect(sections).toEqual(allLessonReviewSections);
-    });
-
-    it("should not include a section if it has no content", () => {
-      const withoutStarterQuiz = pickAvailableSectionsForLesson(
-        pupilLessonOverviewFixture({
-          starterQuiz: [],
-        }),
-      );
-      const withoutExitQuiz = pickAvailableSectionsForLesson(
-        pupilLessonOverviewFixture({
-          exitQuiz: [],
-        }),
-      );
-      const withoutVideo = pickAvailableSectionsForLesson(
-        pupilLessonOverviewFixture({
-          videoMuxPlaybackId: null,
-        }),
-      );
-
-      expect(withoutStarterQuiz).not.toContain("starter-quiz");
-      expect(withoutExitQuiz).not.toContain("exit-quiz");
-      expect(withoutVideo).not.toContain("video");
-    });
-  });
-
   describe("getStaticPaths", () => {
     it("Should not generate pages at build time", async () => {
       const res = await getStaticPaths();
